simplify, remove lib
